stage 2
Handle JavaScript events and save form data to the local storage.

To pass this stage, you need to implement the concept of saving and loading data to and from the local storage. Once users enter any data into the fields, immediately update this data in the local storage. After reloading the page, you should get data from the local storage and change the input fields values. Pay attention when users visit the website for the first time. In this case, there will be no data in local storage. In that case, all fields should be empty.