stage 3
Description

The main idea behind forms is to make requests to the servers. In this stage, when users submit the form data, you will save it into the local storage and show them on the history page. For that reason, you need to create another HTML file.
Objectives

Create another HTML file that shows the submit history to users.

Each data submission should be represented as a div block with the submit-history-card class. Each div block should contain the following elements:

    <p> element with the card-first-name class. Inside this tag, show the first name from the submitted form;
    <p> element with the card-last-name class. Inside this tag, show the last name from the submitted form;
    <p> element with the card-email class. Inside this tag, show the email from the submitted form.
    <p> element with the card-phone class. Inside this tag, show the phone from the submitted form.
    <p> element with the card-company class. Inside this tag, show the company from the submitted form.
    <p> element with the card-address class. Inside this tag, show the address from the submitted form.
    <button> element with the delete-button class. Clicking on it should do nothing for now.

Both pages should contain the navigation bar with the following elements:

    <a> tag with the form-link ID. When users click on this link, navigate them to the main page with the form. To do so, you should set the href attribute value to the path of the main HTML file;
    <a> tag with the history-link ID. When this link is clicked you should navigate to the history page. To do so you should set the href attribute value to the path to the history HTML file.

Once users visit the history page, get all the history of the submission and create a div block with the submit-history-card class for each submission. Add it to the DOM.

After submitting the main page, clear all the input fields on the form.