stage 4/4
Description

This is the last step, nevertheless, an important one. Let's add the option to delete a submission from the history and synchronize data fields within different opened browser tabs.

To synchronize forms within different tabs, use setInterval() function. You can learn more about this function on javascript.info.
Objectives

On the history page, each submit card should contain a button with the delete-button class. When users click on this button, delete the card related to the button. Also, don't forget to delete this submission from the local storage, so it won't be shown after reloading.

When users open two browser tabs with the HTML form and change data in one of them, the data also should be updated in the other tab. Check every 100 ms whether the data in the local storage is changed; if so, update the data in the form.